:recycle: [#53] More refactoring
