+++ src/archiefvernietigingscomponent/destruction/utils.py
@@ -2,21 +2,22 @@ import logging
 import re
 from typing import Optional
 
+from django.conf import settings
 from django.utils.translation import ugettext_lazy as _
 
 from zds_client import ClientError
 
-from archiefvernietigingscomponent.destruction.models import (
+from archiefvernietigingscomponent.report.utils import get_looptijd
+
+from .constants import ListItemStatus
+from .models import (
     BRONORGANISATIE_TEMPLATE_ELEMENT,
     IDENTIFICATIE_TEMPLATE_ELEMENT,
     UUID_TEMPLATE_ELEMENT,
     ZAC_TEMPLATE_ELEMENTS,
+    DestructionListItem,
 )
-from archiefvernietigingscomponent.destruction.service import (
-    fetch_process_type,
-    fetch_resultaat,
-)
-from archiefvernietigingscomponent.report.utils import get_looptijd
+from .service import fetch_process_type, fetch_resultaat
 
 logger = logging.getLogger(__name__)
 
@@ -67,6 +68,30 @@ def get_zaak_link_for_zaakafhandelcomponent(
             link_template = re.sub(re_pattern, value, link_template)
 
     return link_template
+
+
+def get_zaken_chunks(zaken):
+    return (
+        zaken[pos : pos + settings.ZAKEN_PER_QUERY]
+        for pos in range(0, len(zaken), settings.ZAKEN_PER_QUERY)
+    )
+
+
+def set_zaken_availability(zaken):
+    """check if selected zaken are used in other DLs"""
+    zaak_urls = [zaak["url"] for zaak in zaken]
+    selected_zaken = []
+    for chunk in get_zaken_chunks(zaak_urls):
+        selected_zaken += list(
+            DestructionListItem.objects.filter(
+                status__in=[ListItemStatus.suggested, ListItemStatus.processing]
+            )
+            .filter(zaak__in=chunk)
+            .values_list("zaak", flat=True)
+        )
+
+    for zaak in zaken:
+        zaak["available"] = zaak["url"] not in selected_zaken
 
 
 class ServiceNotConfiguredError(Exception):
